% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ArcIntersection.R
\name{ArcIntersection}
\alias{ArcIntersection}
\title{Point(s) at which two arcs intersect on a sphere}
\usage{
ArcIntersection(longitude_1, latitude_1, longitude_2, latitude_2, longitude_3,
  latitude_3, longitude_4, latitude_4, type = c("arc", "arc"),
  EarthRad = 6367.4447)
}
\arguments{
\item{longitude_1}{Decimalised longitude of first point on first arc.}

\item{latitude_1}{Decimalised latitude of first point on first arc.}

\item{longitude_2}{Decimalised longitude of second point on first arc.}

\item{latitude_2}{Decimalised latitude of second point on first arc.}

\item{longitude_3}{Decimalised longitude of first point on second arc.}

\item{latitude_3}{Decimalised latitude of first point on second arc.}

\item{longitude_4}{Decimalised longitude of second point on second arc.}

\item{latitude_4}{Decimalised latitude of second point on second arc.}

\item{type}{Vector to indicate if elements are arcs or Great Circles.}

\item{EarthRad}{Earth radius in kilometres.}
}
\value{
Matrix of longitude-latitude points at which intersection(s) occur.
}
\description{
Given two arcs returns point(s) at which they intersect on a sphere.
}
\details{
Assumes shortest distance between points describes arc. Can define Great Circle-arc intersection with \code{type} option.
}
\examples{
longitude_1 <- runif(1, -180, 180)
longitude_2 <- runif(1, -180, 180)
longitude_3 <- runif(1, -180, 180)
longitude_4 <- runif(1, -180, 180)
latitude_1 <- runif(1, -90, 90)
latitude_2 <- runif(1, -90, 90)
latitude_3 <- runif(1, -90, 90)
latitude_4 <- runif(1, -90, 90)
ArcIntersection(longitude_1, latitude_1, longitude_2, latitude_2,
  longitude_3, latitude_3, longitude_4, latitude_4, EarthRad = 6367.4447)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

