% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC.R
\name{BC}
\alias{BC}
\title{Biogeographic connectedness}
\usage{
BC(
  taxon_locality_matrix,
  tree = NULL,
  count.nodes = FALSE,
  permute.tree = TRUE,
  bootstrap = FALSE,
  jackknife = FALSE,
  resample.replicates = 1000,
  tree.replicates = 1000,
  k = 1000
)
}
\arguments{
\item{taxon_locality_matrix}{Presence-absence (1-0) matrix of taxa (rows) in localities (columns).}

\item{tree}{Optional time-scaled phylogenetic tree of taxa in presence-absence matrix.}

\item{count.nodes}{Option to count nodes rather than use branch-lengths.}

\item{permute.tree}{Option to permute random trees by shuffling tips.}

\item{bootstrap}{Option to bootstrap occurrences.}

\item{jackknife}{Option to jacknife occurrences.}

\item{resample.replicates}{Number of replicates to use for bootstrapping or jacknifing.}

\item{tree.replicates}{Number of replicates to use if permuting trees.}

\item{k}{Value to use to shorten tree and avoid increasing phylogenetic diversity artefact.}
}
\value{
\item{BC_observed}{Empirical results for the: number of localities (\code{N_localities}), number of taxa (\code{N_taxa}), number of links (\code{N_links}), number of endemics (\code{N_endemics}), mean occurrences per locality (\code{Mean_occurrences_per_locality}), and biogeographic connectedness (\code{Biogeographic_connectedness}).}
\item{BC_resampled}{Output for the resampled (bootstrap or jacknife) data if requested.}
\item{BC_permuted}{Output for the permuted data if requested.}
}
\description{
Biogeographic connectedness (sensu Sidor et al.).
}
\details{
Nothing yet.
}
\examples{

# Nothing yet.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com} and Richard J. Butler \email{butler.richard.j@gmail.com}
}
