% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BearingBetweenTwoLongLatPoints.R
\name{BearingBetweenTwoLongLatPoints}
\alias{BearingBetweenTwoLongLatPoints}
\title{Returns bearing from a set of coordinates}
\usage{
BearingBetweenTwoLongLatPoints(start_long, start_lat, end_long, end_lat)
}
\arguments{
\item{start_long}{Decimalised longitude for start point.}

\item{start_lat}{Decimalised latitude for start point.}

\item{end_long}{Decimalised longitude for end point.}

\item{end_lat}{Decimalised latitude for end point.}
}
\value{
Bearing in degrees (0 or 360 for North, 90 for East, etc.).
}
\description{
Returns bearing in degrees for starting and ending longitude-latitude points
}
\details{
Nothing yet.
}
\examples{
# Should return bearing of North (0):
BearingBetweenTwoLongLatPoints(start_long = 0, start_lat = 0,
  end_long = 0, end_lat = 1)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
