% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ChordLengthFromTheta.R
\name{ChordLengthFromTheta}
\alias{ChordLengthFromTheta}
\title{Returns chord length from theta}
\usage{
ChordLengthFromTheta(theta, EarthRad = 6367.4447)
}
\arguments{
\item{theta}{Theta in radians.}

\item{EarthRad}{Radius of the Earth in kilometres.}
}
\value{
Chord length in kilometres.
}
\description{
Returns chord length in kilometres from theta in radians
}
\details{
Nothing yet.
}
\examples{
ChordLengthFromTheta(0.5)
}

