% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ColliderReverser.R
\name{ColliderReverser}
\alias{ColliderReverser}
\title{Reverses a time step to coliision point}
\usage{
ColliderReverser(min_separation, continent_1_longitude_t0,
  continent_1_latitude_t0, continent_1_longitude_t1, continent_1_latitude_t1,
  continent_2_longitude_t0, continent_2_latitude_t0, continent_2_longitude_t1,
  continent_2_latitude_t1, continent_1_euler_longitude,
  continent_1_euler_latitude, continent_2_euler_longitude,
  continent_2_euler_latitude, continent_1_degrees_per_step,
  continent_2_degrees_per_step, EarthRad = 6367.4447, Warn = TRUE)
}
\arguments{
\item{min_separation}{The minimum separation between continents in kilometres.}

\item{continent_1_longitude_t0}{Decimalised longitude of first continent at t0.}

\item{continent_1_latitude_t0}{Decimalised latitude of first continent at t0.}

\item{continent_1_longitude_t1}{Decimalised longitude of first continent at t1.}

\item{continent_1_latitude_t1}{Decimalised latitude of first continent at t1.}

\item{continent_2_longitude_t0}{Decimalised longitude of second continent at t0.}

\item{continent_2_latitude_t0}{Decimalised latitude of second continent at t0.}

\item{continent_2_longitude_t1}{Decimalised longitude of second continent at t1.}

\item{continent_2_latitude_t1}{Decimalised latitude of second continent at t1.}

\item{continent_1_euler_longitude}{Decimalised longitude of Euler pole for first continent.}

\item{continent_1_euler_latitude}{Decimalised latitude of Euler pole for first continent.}

\item{continent_2_euler_longitude}{Decimalised longitude of Euler pole for second continent.}

\item{continent_2_euler_latitude}{Decimalised latitude of Euler pole for second continent.}

\item{continent_1_degrees_per_step}{Degrees per step (speed) of first continent.}

\item{continent_2_degrees_per_step}{Degrees per step (speed) of second continent.}

\item{EarthRad}{Radius of the Earth in kilometres.}

\item{Warn}{Whether or not to print warnings.}
}
\value{
Proportion (0 to 1) of time step at which the minimum separation distance collision occurs.
}
\description{
If a collision happens within a time bin establishes the proportion of time elapsed
}
\details{
Nothing yet.
}
\examples{
min_separation <- 500

continent_1_euler_longitude <- 2
continent_1_euler_latitude <- 89

continent_2_euler_longitude <- 4
continent_2_euler_latitude <- -89

continent_1_degrees_per_step <- 1
continent_2_degrees_per_step <- 1

continent_1_longitude_t0 <- 3
continent_1_latitude_t0 <- 0
continent_2_longitude_t0 <- -2
continent_2_latitude_t0 <- 1

continent_1_start_bearing <-
  BearingBetweenTwoLongLatPoints(continent_1_euler_longitude,
  continent_1_euler_latitude, continent_1_longitude_t0, continent_1_latitude_t0)
continent_2_start_bearing <-
  BearingBetweenTwoLongLatPoints(continent_2_euler_longitude,
  continent_2_euler_latitude, continent_2_longitude_t0, continent_2_latitude_t0)
continent_1_euler_distance <-
  GreatCircleDistanceFromLongLat(continent_1_euler_longitude, continent_1_euler_latitude,
  continent_1_longitude_t0, continent_1_latitude_t0)
continent_2_euler_distance <-
  GreatCircleDistanceFromLongLat(continent_2_euler_longitude, continent_2_euler_latitude,
  continent_2_longitude_t0, continent_2_latitude_t0)

continent_1_longitude_t1 <- EndPoint(continent_1_euler_longitude, continent_1_euler_latitude,
  continent_1_start_bearing + continent_1_degrees_per_step, continent_1_euler_distance,
  EarthRad = 6367.4447)$longitude
continent_1_latitude_t1 <- EndPoint(continent_1_euler_longitude, continent_1_euler_latitude,
  continent_1_start_bearing + continent_1_degrees_per_step, continent_1_euler_distance,
  EarthRad = 6367.4447)$latitude
continent_2_longitude_t1 <- EndPoint(continent_2_euler_longitude, continent_2_euler_latitude,
  continent_2_start_bearing + continent_2_degrees_per_step, continent_2_euler_distance,
  EarthRad = 6367.4447)$longitude
continent_2_latitude_t1 <- EndPoint(continent_2_euler_longitude, continent_2_euler_latitude,
  continent_2_start_bearing + continent_2_degrees_per_step, continent_2_euler_distance,
  EarthRad = 6367.4447)$latitude

GreatCircleDistanceFromLongLat(continent_1_longitude_t0, continent_1_latitude_t0,
  continent_2_longitude_t0, continent_2_latitude_t0)
GreatCircleDistanceFromLongLat(continent_1_longitude_t1, continent_1_latitude_t1,
  continent_2_longitude_t1, continent_2_latitude_t1)

ColliderReverser(min_separation, continent_1_longitude_t0, continent_1_latitude_t0,
  continent_1_longitude_t1, continent_1_latitude_t1, continent_2_longitude_t0,
  continent_2_latitude_t0, continent_2_longitude_t1, continent_2_latitude_t1,
  continent_1_euler_longitude, continent_1_euler_latitude, continent_2_euler_longitude,
  continent_2_euler_latitude, continent_1_degrees_per_step, continent_2_degrees_per_step)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

