% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ContinentSplitter.R
\name{ContinentSplitter}
\alias{ContinentSplitter}
\title{Split apart joined continents}
\usage{
ContinentSplitter(min_separation, longitudes, latitudes, continent_numbers,
  protected_links = matrix(nrow = 0, ncol = 2), EarthRad = 6367.4447)
}
\arguments{
\item{min_separation}{The minimum separation in kilometres between continents.}

\item{longitudes}{Decimalised longitudes of the continents forming the clump in \code{continent_numbers} order.}

\item{latitudes}{Decimalised latitudes of the continents forming the clump in \code{continent_numbers} order.}

\item{continent_numbers}{A character vector of continent numbers.}

\item{protected_links}{A two-column matrix of protected links (those that cannot be severed).}

\item{EarthRad}{Earth radius in kilometres.}
}
\value{
Vector of two continental clumps left after separation.
}
\description{
Return two separate clumps of continents after a separation has occurred.
}
\details{
Nothing yet.
}
\examples{
# Nothing yet
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

