% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DecimaliseLatitudeLongitude.R
\name{DecimaliseLatitudeLongitude}
\alias{DecimaliseLatitudeLongitude}
\title{Decimalises a longitude or latitude}
\usage{
DecimaliseLatitudeLongitude(degrees, minutes = 0, seconds = 0,
  decimals = 0, direction)
}
\arguments{
\item{degrees}{An integer specifying the longitude or latitiude in degrees.}

\item{minutes}{An integer specifying the longitude or latitiude minutes.}

\item{seconds}{An integer specifying the longitude or latitiude seconds.}

\item{decimals}{An integer specifying the longitude or latitiude decimals.}

\item{direction}{The direction (N = North, E = East, S = South, W = West).}
}
\value{
A decimalised latitude or longitude.
}
\description{
Decimalises a longitude or latitude specified in degrees, minutes, seconds, and decimals
}
\details{
Nothing yet.
}
\examples{
DecimaliseLatitudeLongitude(degrees = 179, minutes = 12,
  seconds = 13, decimals = 67, direction = "W")
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

