\name{dispeRse-package}

\alias{dispeRse-package}

\alias{dispeRse}

\docType{package}

\title{
Models for Paleobiogeography
}

\description{
Functions for simulating and modelling phylogenetic patterns of paleobiogeography. [ADD MORE]
}

\details{
\tabular{ll}{
Package: \tab dispeRse\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2015-04-30\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}

\author{
Laura C. Soul <
Graeme T. Lloyd <graemetlloyd@gmail.com>
}

\references{
Nothing yet.
}

\keyword{ biogeography,phylogeny }

\examples{
# Nothing yet.
}
