% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DivingBoard.R
\name{DivingBoard}
\alias{DivingBoard}
\title{Point where dispersal route hits coast}
\usage{
DivingBoard(start_longitude, start_latitude, end_longitude, end_latitude,
  continent_longitude, continent_latitude, continent_radius,
  EarthRad = 6367.4447)
}
\arguments{
\item{start_longitude}{Decimalised longitude of organism's starting point.}

\item{start_latitude}{Decimalised latitude of organism's starting point.}

\item{end_longitude}{Decimalised longitude of organism's finishing point.}

\item{end_latitude}{Decimalised latitude of organism's finishing point.}

\item{continent_longitude}{Theta in radians.}

\item{continent_latitude}{Theta in radians.}

\item{continent_radius}{Continent radius in kilometres}

\item{EarthRad}{Radius of the Earth in kilometres.}
}
\value{
Decimalised longitude and latitude of collision point and remaining unspent dispersal distance in kilometres.
}
\description{
Finds point at which an organism collides with the coast when dispersing
}
\details{
Nothing yet. Circular continent.
}
\examples{
#DivingBoard()
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

