% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispeRse.R
\name{.to_grid}
\alias{.to_grid}
\title{Convert from geographic coordinates in a given projection system to the
relative position in rows and columns of a grid.}
\usage{
.to_grid(coords, grid)
}
\arguments{
\item{coords}{A SpatialPointsDataFrame. Must contain a column dates with the
start date of the dispersal from each point.}

\item{grid}{A Raster*. The coordinates will be converted to this grid.}
}
\value{
A DataFrame with the converted coordinates.
}
\description{
Convert from geographic coordinates in a given projection system to the
relative position in rows and columns of a grid.
}
