% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EqualAreaRectangularGrid.R
\name{EqualAreaRectangularGrid}
\alias{EqualAreaRectangularGrid}
\title{Equal-area grid from lines of longitude and latitude}
\usage{
EqualAreaRectangularGrid(N_longitude, N_latitude)
}
\arguments{
\item{N_longitude}{Number of East-West bins.}

\item{N_latitude}{Number of North-South bins.}
}
\value{
The decimalised lines of longitude and latitude that describe the grid.
}
\description{
Makes an equal-area grid of specified size from lines of longitude and latitude
}
\details{
Nothing yet.
}
\examples{
# Create a 20-by-10 (200 cell) grid:
grids <- EqualAreaRectangularGrid(N_longitude = 20, N_latitude = 10)

# View the output:
grids

# Visualise the output using a Mercator projection:
plot(0, 0, type = "n", xlim = c(-180, 180), ylim = c(-90, 90),
  xlab = "Longitude", ylab = "Latitude")
for(i in 1:length(grids$latitude_breaks))
  lines(c(-180, 180), c(grids$latitude_breaks[i], grids$latitude_breaks[i]), lty = 2)
for(i in 1:length(grids$longitude_breaks))
  lines(c(grids$longitude_breaks[i], grids$longitude_breaks[i]), c(-90, 90), lty = 2)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

