% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispeRse.R
\docType{data}
\name{euro_dates}
\alias{euro_dates}
\title{Coordinates and earliest dates (med cal BP) for European Neolithic sites.
Dataset adapted from the supplementary information in Pinhasi et al. 2005
(https://doi.org/10.1371/journal.pbio.0030410).}
\format{
A SpatialPointsDataFrame object.
}
\usage{
euro_dates
}
\description{
Coordinates and earliest dates (med cal BP) for European Neolithic sites.
Dataset adapted from the supplementary information in Pinhasi et al. 2005
(https://doi.org/10.1371/journal.pbio.0030410).
}
\keyword{datasets}
