% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispeRse.R
\docType{data}
\name{euro_npp}
\alias{euro_npp}
\title{Transformed Net Primary Production (NPP) from 11 ka to 4 ka at 1000 yr steps.
Calculated with the Miami formula using paleoclimatic data from Beyer et al. 2020.
(https://doi.org/10.1038/s41597-020-0552-1). Clipped to max=1350, squared and
scaled to [0,1].}
\format{
A RasterStack object.
}
\usage{
euro_npp
}
\description{
Transformed Net Primary Production (NPP) from 11 ka to 4 ka at 1000 yr steps.
Calculated with the Miami formula using paleoclimatic data from Beyer et al. 2020.
(https://doi.org/10.1038/s41597-020-0552-1). Clipped to max=1350, squared and
scaled to [0,1].
}
\keyword{datasets}
