% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispeRse.R
\docType{data}
\name{euro_terr}
\alias{euro_terr}
\title{Reclassified terrain layer with elevation > 1750 m as barriers and rivers
and coastline as corridors. Terrain reclassified from SRTM. Rivers rasterized
from GSHHG (https://www.soest.hawaii.edu/pwessel/gshhg/) and coastline
rasterized from rnaturalearth.}
\format{
A RasterLayer object.
}
\usage{
euro_terr
}
\description{
Reclassified terrain layer with elevation > 1750 m as barriers and rivers
and coastline as corridors. Terrain reclassified from SRTM. Rivers rasterized
from GSHHG (https://www.soest.hawaii.edu/pwessel/gshhg/) and coastline
rasterized from rnaturalearth.
}
\keyword{datasets}
