% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EverythingFunction.R
\name{EverythingFunction}
\alias{EverythingFunction}
\title{Global function to simulate evolution on a sphere}
\usage{
EverythingFunction(N_steps = 1000, N_continents = 7, radius = 2000,
  start_configuration = "supercontinent", squishiness = 0.25,
  stickiness = 0.95, continent_speed_mean = 500, continent_speed_sd = 250,
  EarthRad = 6367.4447, polar = FALSE)
}
\arguments{
\item{N_steps}{The number of time steps in the simulation.}

\item{N_continents}{The (maximum) number of individual continents.}

\item{radius}{The radius of each circular continent.}

\item{start_configuration}{One of "random separate", "random overlap", "supercontinent", or "max separate".}

\item{squishiness}{A value from 0 (continents can never overlap) to 1 (continents can overlap completely).}

\item{stickiness}{Probability of two conjoined continents remaining conjoined in the next time step.}

\item{continent_speed_mean}{Mean continent speed (kilometers per time step).}

\item{continent_speed_sd}{Standard deviation of continent speed (kilometers per time step).}

\item{EarthRad}{Radius of the Earth in kilometres.}
}
\value{
A magic table of awesomeness.
}
\description{
Global function to generate simulated continents and clades on a sphere
}
\details{
Nothing yet.
}
\examples{
EverythingFunction <- function(N_steps = 1000, N_continents = 7, radius = 2000,
  start_configuration = "supercontinent", squishiness = 0.25, stickiness = 0.95,
  continent_speed_mean = 500, continent_speed_sd = 250, EarthRad = 6367.4447)
}

