% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GreatCircleDistanceFromTheta.R
\name{GreatCircleDistanceFromTheta}
\alias{GreatCircleDistanceFromTheta}
\title{Returns great circle distance from theta}
\usage{
GreatCircleDistanceFromTheta(theta, EarthRad = 6367.4447)
}
\arguments{
\item{theta}{Theta in radians.}

\item{EarthRad}{Radius of the Earth in kilometres.}
}
\value{
Great circle distance in kilometres.
}
\description{
Returns great circle distance in kilometres from theta in radians
}
\details{
Nothing yet.
}
\examples{
GreatCircleDistanceFromTheta(0.5)
}

