% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GreatCircleDistanceMatrix.R
\name{GreatCircleDistanceMatrix}
\alias{GreatCircleDistanceMatrix}
\title{Great Circle Distance Matrix Between Coordinates}
\usage{
GreatCircleDistanceMatrix(longs, lats, EarthRad = 6367.4447)
}
\arguments{
\item{longs}{A vector of decimalised longitudes.}

\item{lats}{A vector of decimalised latitudes.}

\item{EarthRad}{Radius of the Earth in kilometres.}
}
\value{
A distancematrix of minimum Great Circle Distances.
}
\description{
Returns a distance matrix between points of longitude and latitude
}
\details{
Nothing yet.
}
\examples{
GreatCircleDistanceMatrix(c(-179, 0, 179), c(89, 0, -89))
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

