% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreatCircleIntersection.R
\name{GreatCircleIntersection}
\alias{GreatCircleIntersection}
\title{Point(s) at which two Great Circles intersect on a sphere}
\usage{
GreatCircleIntersection(
  longitude_1,
  latitude_1,
  longitude_2,
  latitude_2,
  longitude_3,
  latitude_3,
  longitude_4,
  latitude_4,
  EarthRad = 6367.4447
)
}
\arguments{
\item{longitude_1}{Decimalised longitude of first point on first Great Circle.}

\item{latitude_1}{Decimalised latitude of first point on first Great Circle.}

\item{longitude_2}{Decimalised longitude of second point on first Great Circle.}

\item{latitude_2}{Decimalised latitude of second point on first Great Circle.}

\item{longitude_3}{Decimalised longitude of first point on second Great Circle.}

\item{latitude_3}{Decimalised latitude of first point on second Great Circle.}

\item{longitude_4}{Decimalised longitude of second point on second Great Circle.}

\item{latitude_4}{Decimalised latitude of second point on second Great Circle.}

\item{EarthRad}{Earth radius in kilometres.}
}
\value{
Matrix of longitude-latitude points at which intersection(s) occur.
}
\description{
Given two Great Circles returns point(s) at which they intersect on a sphere.
}
\details{
Nothing yet.
}
\examples{
longitude_1 <- stats::runif(1, -180, 180)
longitude_2 <- stats::runif(1, -180, 180)
longitude_3 <- stats::runif(1, -180, 180)
longitude_4 <- stats::runif(1, -180, 180)
latitude_1 <- stats::runif(1, -90, 90)
latitude_2 <- stats::runif(1, -90, 90)
latitude_3 <- stats::runif(1, -90, 90)
latitude_4 <- stats::runif(1, -90, 90)
GreatCircleIntersection(longitude_1, latitude_1, longitude_2, latitude_2,
  longitude_3, latitude_3, longitude_4, latitude_4)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
