% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HowManySeparateContinents.R
\name{HowManySeparateContinents}
\alias{HowManySeparateContinents}
\title{How many separate continents are there}
\usage{
HowManySeparateContinents(min_separation, longitudes, latitudes,
  EarthRad = 6367.4447)
}
\arguments{
\item{min_separation}{Minimum separation allowed between continental centres in kilometres.}

\item{longitudes}{A vector of the decimalised longitudes for each continental centre.}

\item{latitudes}{A vector of the decimalised latitudes for each continental centre.}

\item{EarthRad}{Earth radius in kilometres.}
}
\value{
A vector of length equal to the number of separate continents.
}
\description{
Given a minimum separation between continental centres records how many separate continents exist
}
\details{
Nothing yet.
}
\examples{
HowManySeparateContinents(100, c(179, -179), c(-89, 89))
}

