% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IntercontinentalLinks.R
\name{IntercontinentalLinks}
\alias{IntercontinentalLinks}
\title{Which continents are linked?}
\usage{
IntercontinentalLinks(min_separation, longitudes, latitudes,
  EarthRad = 6367.4447)
}
\arguments{
\item{min_separation}{Minimum separation allowed between continental centres in kilometres.}

\item{longitudes}{A vector of the decimalised longitudes for each continental centre.}

\item{latitudes}{A vector of the decimalised latitudes for each continental centre.}

\item{EarthRad}{Earth radius in kilometres.}
}
\value{
A square matrix with intercontinental links scored as ones.
}
\description{
Returns a continent-by-continent matrix with intercontinental links scored as ones
}
\details{
Nothing yet.
}
\examples{
IntercontinentalLinks(0, c(179, -179), c(-89, 89))
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

