% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MagicPortal.R
\name{MagicPortal}
\alias{MagicPortal}
\title{Sweepstakes dispersal}
\usage{
MagicPortal(start_longitude, start_latitude, end_longitude, end_latitude,
  start_continent, touching_continents, continent_centres, continent_radius,
  EarthRad = 6367.4447)
}
\arguments{
\item{start_longitude}{Decimalised longitude of organism's dispersal starting point.}

\item{start_latitude}{Decimalised latitude of organism's dispersal starting point.}

\item{end_longitude}{Decimalised longitude of organism's dispersal ending point.}

\item{end_latitude}{Decimalised latitude of organism's dispersal ending point.}

\item{start_continent}{Number of starting continent.}

\item{touching_continents}{Character vector of continental clusters separated by ampersands.}

\item{continent_centres}{A two-column (longitude, latitude) of the centres of each continent in order (1 at top to N at bottom).}

\item{continent_radius}{The radius of the continent in kilometres.}

\item{EarthRad}{Earth radius in kilometres.}
}
\value{
The new position of the organism following a sweepstakes dispersal.
}
\description{
If sweeptakes dispersal occurs picks a landing spot.
}
\examples{
#MagicPortal()
}
\author{
Laura C. Soul \email{lauracsoul@gmail.com} and Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

