% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MapContinents.R
\name{MapContinents}
\alias{MapContinents}
\title{Maps continents on Mercator projection}
\usage{
MapContinents(continent_centres, radius, xlim = c(-180, 180), ylim = c(-90,
  90), resolution = 1000, sea_colour = "blue", land_colour = "green", ...)
}
\arguments{
\item{continent_centres}{A two-column matrix of decimalised longitudes (first column) and latitudes (second column).}

\item{radius}{A single radius value for each continent in kilometres.}

\item{xlim}{A single radius value for each continent in kilometres.}

\item{ylim}{A single radius value for each continent in kilometres.}

\item{resolution}{The number of points to use to plot each continent.}

\item{sea_colour}{The colour value to use to plot the sea.}

\item{land_colour}{The colour value used to plot the land.}

\item{...}{Other arguments to be passed to \link{plot}.}
}
\value{
A Mercator projection plot of each continent.
}
\description{
Given a set of continent centres and their radius plots them on a Mercator projection
}
\details{
Nothing yet.
}
\examples{
MapContinents(continent_centres = matrix(c(-0, 0), ncol=2), radius = 2000)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

