% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/One2ManyGreatCircleDistance.R
\name{One2ManyGreatCircleDistance}
\alias{One2ManyGreatCircleDistance}
\title{One-to-many Great Circle Distances}
\usage{
One2ManyGreatCircleDistance(point_longitude, point_latitude, point_longitudes,
  point_latitudes, EarthRad = 6367.4447)
}
\arguments{
\item{point_longitude}{Decimalised longitude of single point.}

\item{point_latitude}{Decimalised latitude of single point.}

\item{point_longitudes}{Decimalised longitudes of many points.}

\item{point_latitudes}{Decimalised latitudes of many points.}

\item{EarthRad}{Radius of the Earth in kilometres.}
}
\value{
A vector of minimum Great Circle Distances between single point and each of many points.
}
\description{
Returns a set of distances from a single point of longitude and latitude
}
\details{
Nothing yet.
}
\examples{
One2ManyGreatCircleDistance(0, 0, c(-179, 0, 179), c(89, 0, -89))
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

