% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OpenSpots.R
\name{OpenSpots}
\alias{OpenSpots}
\title{Find open spots for next continent in a supercontinent}
\usage{
OpenSpots(longitudes, latitudes, min_separation, EarthRad = 6367.4447)
}
\arguments{
\item{longitudes}{Decimalised longitudes of existing continents.}

\item{latitudes}{Decimalised longitudes of existing continents.}

\item{min_separation}{The minimum separation between continents in kilomeres.}

\item{EarthRad}{Earth radius in kilometres.}
}
\value{
A two-column (longitude and latitude) matrix of open spots.
}
\description{
Takes position of existing continents and finds the closest possible position(s) for a new continent
}
\examples{
OpenSpots(c(0, 1), c(0, 0), 111.132874953663, EarthRad = 6367.4447)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

