% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispeRse.R
\docType{data}
\name{ppnb}
\alias{ppnb}
\title{Coordinates and earliest dates (med cal BP) for Late Pre-Pottery Neolithic B
sites in the Near East. Dataset adapted from the supplementary information
in Pinhasi et al. 2005 (https://doi.org/10.1371/journal.pbio.0030410).}
\format{
A DataFrame object.
}
\usage{
ppnb
}
\description{
Coordinates and earliest dates (med cal BP) for Late Pre-Pottery Neolithic B
sites in the Near East. Dataset adapted from the supplementary information
in Pinhasi et al. 2005 (https://doi.org/10.1371/journal.pbio.0030410).
}
\keyword{datasets}
