% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RandomBearingDraw.R
\name{RandomBearingDraw}
\alias{RandomBearingDraw}
\title{Global function to simulate evolution on a sphere}
\usage{
RandomBearingDraw(n = 1, shape_parameter = 1)
}
\arguments{
\item{N_steps}{The number of time steps in the simulation.}

\item{N_continents}{The (maximum) number of individual continents.}

\item{radius}{The radius of each circular continent.}

\item{start_configuration}{One of "random separate", "random overlap", "supercontinent", or "max separate".}

\item{squishiness}{A value from 0 (continents can never overlap) to 1 (continents can overlap completely).}

\item{stickiness}{Probability of two conjoined continents remaining conjoined in the next time step.}

\item{continent_speed_mean}{Mean continent speed (kilometers per time step).}

\item{continent_speed_sd}{Standard deviation of continent speed (kilometers per time step).}

\item{EarthRad}{Radius of the Earth in kilometres.}

\item{polar}{Whether continents are allowed to exist exactly at the poles.}
}
\value{
A magic table of awesomeness.
}
\description{
Global function to generate simulated continents and clades on a sphere
}
\details{
Nothing yet.
}
\examples{
# Set seed for example:
set.seed(21)

# Unbiased random draw (N-S and E-W draws equally likely):
hist(RandomBearingDraw(n = 10000, shape_parameter = 1), xlim=c(0, 360),
  breaks=36, xlab="Bearing (degrees)")

# Random draw biased towards E-W draws:
hist(RandomBearingDraw(n = 10000, shape_parameter = 10), xlim=c(0, 360),
  breaks=36,xlab="Bearing (degrees)")

# Random draw biased towards N-S:
hist(RandomBearingDraw(n = 10000, shape_parameter = 0.1), xlim=c(0, 360),
  breaks=36, xlab="Bearing (degrees)")
}

