% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispeRse.R
\name{run_disp}
\alias{run_disp}
\title{Lorem ipsum dolor.}
\usage{
run_disp(environment, terrain, r, phi, coords, iter, t)
}
\arguments{
\item{environment}{lorem}

\item{terrain}{lorem}

\item{r}{lorem}

\item{phi}{lorem}

\item{coords}{lorem}

\item{iter}{lorem}

\item{t}{lorem}
}
\value{
lorem
}
\description{
Lorem ipsum dolor.
}
