% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispeRse.R
\name{sim_dispersal}
\alias{sim_dispersal}
\title{Lorem ipsum dolor.}
\usage{
sim_dispersal(start_coords, start_dates, environment, base_speed = 1)
}
\arguments{
\item{start_coords}{lorem}

\item{start_dates}{lorem}

\item{environment}{lorem}

\item{base_speed}{lorem}
}
\value{
lorem
}
\description{
Lorem ipsum dolor.
}
