% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispeRse.R
\name{simulate_dispersal}
\alias{simulate_dispersal}
\title{Simulates first arrival times from one or more origins.}
\usage{
simulate_dispersal(
  environment,
  terrain,
  coords,
  num_iter,
  r = 0.025,
  phi = 0.5,
  t = 25,
  dist = 50,
  accel = 3,
  gamma = 1
)
}
\arguments{
\item{environment}{A RasterLayer. The local K and r depend on its values,
which preferably range from 0 to 1 (fraction of maximum K).}

\item{terrain}{A RasterLayer. Values must be 0 = no effect, 1 = barrier,
2 = corridor.}

\item{coords}{A DataFrame. Must contain columns x, y, and date with the
coordinates and starting date (yr BP) of each origin. Coordinates must be in
the same system as the environment and terrain layers.}

\item{num_iter}{Numeric. Number of iterations (steps) of the model.}

\item{r}{Numeric. The annual growth rate as a fraction.}

\item{phi}{Numeric. The fission threshold as a fraction of carrying capacity.}

\item{t}{Numeric. The duration, in years, of a generation (model time step).}

\item{dist}{Numeric. The distance, in km, that migrants move over a
generation.}

\item{accel}{Numeric. The factor by which the usual distance is increased
along corridors. E.g. if dist = 50 km and accel = 3, migrants can move up to
150 km along a corridor.}

\item{gamma}{Numeric. A power that controls the shape of the dependency
between r and the environment.}
}
\value{
A RasterLayer with simulated arrival times.
}
\description{
The model uses density-dependent growth and emigration. The carrying
capacity, growth rates and mobility are allowed to vary with the environment.
}
\details{
The simulation starts with n populated cells at coordinates and start times
defined by the parameter coords and runs for a number of time. Each time step
corresponds to a generation (defined by parameter t). Growth is applied to
every populated cell using a logistic model and emigration to a neighboring
cell is calculated from an asymptotic threshold model.

The carrying capacity (K) in the density-dependend growth and emigration
models is determined by an environment raster, which normally represents a
variable assumed to affect population density (e.g. net primary production)
scaled to 0-1 range.

While carrying capacity depends linearly on the environment, the dependence
of the growth rate is allowed to be controlled by a power gamma.

For emigration, a threshold phi is considered, expressed as a fraction of
carrying capacity. Migrants move to the cell with the highest environmental
value in their Moore neighborhood (8 cells).

Terrain can be represented by a raster specifying barriers (e.g. mountains),
which block movement, and corridors (e.g. rivers), which accelerate movement.
For the latter, an acceleration factor can be specified.
}
