% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SphericalAngleForEquilateralTriangleFromGreatCircleSideLength.R
\name{SphericalAngleForEquilateralTriangleFromGreatCircleSideLength}
\alias{SphericalAngleForEquilateralTriangleFromGreatCircleSideLength}
\title{Finds the spherical angle for an equilateral triangle}
\usage{
SphericalAngleForEquilateralTriangleFromGreatCircleSideLength(side_length,
  EarthRad = 6367.4447)
}
\arguments{
\item{side_length}{The great circle distance of a side of the triangle in kilometres.}

\item{EarthRad}{Radius of the Earth in kilometres.}
}
\value{
Spherical angle in degrees.
}
\description{
Returns the spherical angle in degrees for an equilateral triangle of known side length
}
\details{
Nothing yet.
}
\examples{
SphericalAngleForEquilateralTriangleFromGreatCircleSideLength(1000)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

