% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SphericalCapArea.R
\name{SphericalCapArea}
\alias{SphericalCapArea}
\title{Spherical Cap Area}
\usage{
SphericalCapArea(radius, EarthRad = 6367.4447)
}
\arguments{
\item{radius}{Radius of the cap in kilometres.}

\item{EarthRad}{Radius of the Earth in kilometres.}
}
\value{
The surface area of the cap in square-kilometres.
}
\description{
Returns the surface area of a cap (circle) on a sphere
}
\details{
Nothing yet.
}
\examples{
SphericalCapArea(2000)
}

