% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StartingPoints.R
\name{StartingPoints}
\alias{StartingPoints}
\title{Generate a matrix of longitudes and latitudes of continental centres}
\usage{
StartingPoints(N_continents = 7, radius = 2000,
  start_configuration = "supercontinent", squishiness = 0.25,
  EarthRad = 6367.4447, polar = FALSE)
}
\arguments{
\item{N_continents}{The (maximum) number of individual continents}

\item{radius}{The radius of each circular continent.}

\item{start_configuration}{One of "random separate", "random overlap", "supercontinent", or "max separate".}

\item{squishiness}{A value from 0 (continents can never overlap) to 1 (continents can overlap completely)}

\item{EarthRad}{Eartn radius in kilometres.}

\item{polar}{TRUE/FALSE Is there a continent starting on the south pole?}
}
\value{
A matrix of longitudes and latitudes describing the centres of circular continents.
}
\description{
Generates a matrix of longitudes and latitudes describing the centres of circular continents on a sphere
}
\details{
Nothing yet.
}
\examples{
StartingPoints(N_continents = 7, radius = 2000,
   start_configuration = "supercontinent", squishiness = 0.1,
   EarthRad = 6367.4447, polar=FALSE)
}

