% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThetaFromChordLength.R
\name{ThetaFromChordLength}
\alias{ThetaFromChordLength}
\title{Get theta from chord length}
\usage{
ThetaFromChordLength(chord_length, EarthRad = 6367.4447)
}
\arguments{
\item{chord_length}{The chord length in kilometres.}

\item{EarthRad}{Radius of the Earth in kilometres.}
}
\value{
Theta in radians.
}
\description{
Returns theta in radians from a chord length in kilometres
}
\details{
Nothing yet.
}
\examples{
ThetaFromChordLength(1000)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
