% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TreeWalkerContinuous.R
\name{TreeWalkerContinuous}
\alias{TreeWalkerContinuous}
\title{Brownian motion (random walk) on a sphere}
\usage{
TreeWalkerContinuous(tree, slon = 0, slat = 0, niter = 1000,
  steplengthmean = 0, steplengthsd = 100, EarthRad = 6367.4447)
}
\arguments{
\item{tree}{tree (as phylo object) to use}

\item{slon}{starting longitude}

\item{slat}{starting latitude}

\item{niter}{number of time steps to use}

\item{steplengthmean}{mean used for random walk draws}

\item{steplengthsd}{standard deviation used for random walk draws}

\item{EarthRad}{Earth radius in kilometres.}
}
\value{
A list of edges with matrices showing times and coordinates
}
\description{
Perform Brownian motion (a random walk) on a sphere.
}
\examples{
tree <- rtree(10)
TreeWalkerContinuous(tree, slat = 0, slon = 0, niter = 1000, steplengthmean = 0, steplengthsd = 100)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\keyword{random}
\keyword{walk}

