% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WhichSupercontinent.R
\name{WhichSupercontinent}
\alias{WhichSupercontinent}
\title{Which supercontinent?}
\usage{
WhichSupercontinent(continent, supercontinents)
}
\arguments{
\item{continent}{Continent number desired.}

\item{supercontinents}{Contienntla cluster vector.}
}
\value{
Index of \code{continent} in \code{supercontinents}.
}
\description{
Which continental cluster does the current continent belong to.
}
\details{
Nothing yet.
}
\examples{
WhichSupercontinent("3", c("1&2", "3&4"))
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

