% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalSimulator.R
\name{DispersalSimulator}
\alias{DispersalSimulator}
\title{Global function to simulate evolution on a sphere}
\usage{
DispersalSimulator(
  N_steps = 1000,
  organism_multiplier = 5,
  N_continents = 7,
  radius = 2000,
  start_configuration = "random separate",
  squishiness = 0.25,
  stickiness = 0.95,
  continent_speed_mean = 5,
  continent_speed_sd = 2,
  organism_step_sd = 100,
  organism_step_shape = 1,
  b = 0.1,
  d = 0.05,
  sweepstakes = 0,
  EarthRad = 6367.4447,
  polar = FALSE
)
}
\arguments{
\item{N_steps}{The number of time steps in the simulation.}

\item{organism_multiplier}{The number of organism time steps to be taken per continent time step.}

\item{N_continents}{The (maximum) number of individual continents.}

\item{radius}{The radius of each circular continent.}

\item{start_configuration}{One of "random separate", "random overlap", "supercontinent", or "max separate".}

\item{squishiness}{A value from 0 (continents can never overlap) to 1 (continents can overlap completely).}

\item{stickiness}{Probability of two conjoined continents remaining conjoined in the next time step.}

\item{continent_speed_mean}{Mean continent speed (kilometers per time step).}

\item{continent_speed_sd}{Standard deviation of continent speed (kilometers per time step).}

\item{organism_step_sd}{Standard deviation used for random walk draws for organisms.}

\item{organism_step_shape}{Shape parameter to use for drawing bearings for organisms.}

\item{b}{Per-lineage birth (speciation) rate.}

\item{d}{Per-lineage death (extinction) rate.}

\item{sweepstakes}{The probability of a sweepstakes dispersal occuring each time an organism hits the coast.}

\item{EarthRad}{Radius of the Earth in kilometres.}

\item{polar}{Whether continents are allowed to exist exactly at the poles.}
}
\value{
A magic table of awesomeness.
}
\description{
Global function to generate simulated continents and clades on a sphere
}
\details{
Nothing yet.
}
\examples{
#DispersalSimulator(N_steps = 100, organism_multiplier = 5, N_continents = 2, radius = 2000,
#  start_configuration = "random separate", squishiness = 0.25, stickiness = 0.95,
#  continent_speed_mean = 5, continent_speed_sd = 2, organism_step_sd = 100, sweepstakes = 0, 
#  b = 0.1, d = 0.05, EarthRad = 6367.4447,  polar = FALSE)
}
\author{
Laura C. Soul \email{lauracsoul@gmail.com} and Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
