% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EndPoint.R
\name{EndPoint}
\alias{EndPoint}
\title{Random walk on a sphere}
\usage{
EndPoint(slat = 0, slong = 0, bearing = 0, distance = 1)
}
\arguments{
\item{bearing}{heading in degrees from north}

\item{distance}{length of step in km}

\item{latitude}{starting latitude}

\item{longitude}{starting longitude}
}
\value{
lat final latitude

long final longitude

bearing final bearing in degrees from north
}
\description{
This function allows you calculate the final coordinates and bearing after one step of a random walk on a sphere.
}
\examples{
EndPoint(slat=0, slong=0, bearing=90, distance=111)
}
\keyword{random}
\keyword{walk}

