% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/EndPoint.R
\name{EndPoint}
\alias{EndPoint}
\title{Random walk on a sphere}
\usage{
EndPoint(slong = 0, slat = 0, bearing = 0, distance = 1)
}
\arguments{
\item{slong}{starting longitude}

\item{slat}{starting latitude}

\item{bearing}{heading in degrees from north}

\item{distance}{length of step in km}
}
\value{
lat final latitude

long final longitude

bearing final bearing in degrees from north
}
\description{
This function allows you calculate the final coordinates and bearing after one step of a random walk on a sphere.
}
\examples{
EndPoint(slong = 0, slat = 0, bearing = 90, distance = 111)
}
\keyword{random}
\keyword{walk}

