% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EqualAreaRectangularGrid.R
\name{EqualAreaRectangularGrid}
\alias{EqualAreaRectangularGrid}
\title{Equal-area grid from lines of longitude and latitude}
\usage{
EqualAreaRectangularGrid(N_latitude, N_longitude)
}
\arguments{
\item{N_latitude}{Radius of the Earth in kilometres.}

\item{N_longitude}{Theta in radians.}
}
\value{
The lines of longitudde and latitude that describe the grid.
}
\description{
Makes an equal-area grid of specified size from lines of longitude and latitude
}
\details{
Nothing yet.
}
\examples{
EqualAreaRectangularGrid(N_latitude = 10, N_longitude = 20)
# map results
# Mercator projection
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

