% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GreatCircleDistanceFromLongLat.R
\name{GreatCircleDistanceFromLongLat}
\alias{GreatCircleDistanceFromLongLat}
\title{Great circle distance from a set of coordinates}
\usage{
GreatCircleDistanceFromLongLat(long1, lat1, long2, lat2, EarthRad = 6367.4447,
  Warn = TRUE)
}
\arguments{
\item{long1}{Decimalised longitude for first point.}

\item{lat1}{Decimalised latitude for first point.}

\item{long2}{Decimalised longitude for second point.}

\item{lat2}{Decimalised latitude for second point.}

\item{EarthRad}{Radius of the Earth in kilometres.}

\item{Warn}{Whether or not to print warnings.}
}
\value{
Great circle distance in kilometres.
}
\description{
Returns great circle distance in kilometres from a two longitude-latitude points
}
\details{
Nothing yet.
}
\examples{
# Return one-degree of latitude (approximately 111 kilometres):
GreatCircleDistanceFromLongLat(0, 0, 0, 1)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

