% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MoveTogether.R
\name{MoveTogether}
\alias{MoveTogether}
\title{Move multiple 'continents' about a Euler pole}
\usage{
MoveTogether(centers, pole, angle, EarthRad = 6367.4447)
}
\arguments{
\item{centers}{A matix with columns corresponding to continent central positions in degrees. Row 1 for longitude row 2 for latitude. Or a long lat vector for a single continent.}

\item{pole}{A vector of the long lat of the pole of rotation}

\item{angle}{The angle of rotation about the pole in degrees}

\item{EarthRad}{Earth radius in kilometres.}
}
\value{
a matix of the new long lat in degrees of the centers of each continent after moving
}
\description{
Moves multiple 'continents' (which are temporarily or permanently joined) by their midpoints keeping them exactly the same distance from one another so that they behave as a single unit
}
\examples{
centers<-rbind(c(40, -50, -170), c(34, -60, 79))
pole<-c(90, -20)
MoveTogether(centers, pole, angle = 5)
}
\author{
Laura C. Soul \email{lauracsoul@gmail.com}
}
\keyword{rotate}
\keyword{sphere}

