% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/One2ManyGreatCircleDistance.R
\name{One2ManyGreatCircleDistance}
\alias{One2ManyGreatCircleDistance}
\title{One-to-many Great Circle Distances}
\usage{
One2ManyGreatCircleDistance(point_longitude, point_latitude, point_longitudes,
  point_latitudes, EarthRad = 6367.4447)
}
\arguments{
\item{EarthRad}{Radius of the Earth in kilometres.}

\item{longs}{A vector of decimalised longitudes.}

\item{lats}{A vector of decimalised latitudes.}
}
\value{
A distancematrix of minimum Great Circle Distances.
}
\description{
Returns a set of distances distance matrix between points of longitude and latitude
}
\details{
Nothing yet.
}
\examples{
One2ManyGreatCircleDistance(0, 0, c(-179, 0, 179), c(89, 0, -89))
}

