% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomBearingDraw.R
\name{RandomBearingDraw}
\alias{RandomBearingDraw}
\title{Make random draws for bearings}
\usage{
RandomBearingDraw(n = 1, shape_parameter = 1)
}
\arguments{
\item{n}{The number of randomly drawn bearings required.}

\item{shape_parameter}{The shape parameter used as both values in a beta distribution.}
}
\value{
A vector of bearing(s) of length n.
}
\description{
Make random draws for bearings allowing bias towards North-South or East-West
}
\details{
Nothing yet.
}
\examples{
# Set seed for example:
set.seed(21)

# Unbiased random draw (N-S and E-W draws equally likely):
hist(RandomBearingDraw(n = 10000, shape_parameter = 1), xlim=c(0, 360),
  breaks=36, xlab="Bearing (degrees)")

# Random draw biased towards E-W draws:
hist(RandomBearingDraw(n = 10000, shape_parameter = 10), xlim=c(0, 360),
  breaks=36,xlab="Bearing (degrees)")

# Random draw biased towards N-S:
hist(RandomBearingDraw(n = 10000, shape_parameter = 0.1), xlim=c(0, 360),
  breaks=36, xlab="Bearing (degrees)")
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
